example-api/requirements.txt
Flask==1.0.2
python-dateutil

# testing
pytest